Find_Odd_Occurance_XOR.cpp
The C++ program finds the number with an odd occurrence in an array using the bitwise XOR operator. By XORing all elements in the array, the program effectively cancels out pairs of the same numbers, leaving behind the element with an odd occurrence. This bitwise operation enables a compact and efficient solution for identifying the number that occurs an odd number of times in the array.